% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrowthModels.R
\name{GrowthModels}
\alias{GrowthModels}
\title{Apply Stand-level Growth Models from Maleki et al. (2022)}
\usage{
GrowthModels(Data, ModelsAndParameters, nSpecies)
}
\arguments{
\item{Data}{A \code{data.table} object representing the forest inventory to which the
Maleki et al. (2022) growth models will be applied.}

\item{ModelsAndParameters}{A list structured according to the \code{ModelsAndParameters}
dataset documentation, containing growth models and parameters from Maleki et al. (2022).}

\item{nSpecies}{A numeric vector specifying the species codes for which the growth models
should be applied, corresponding to species codes in \code{Data}.}
}
\value{
Invisibly returns the modified \code{Data} table with updated growth projections.
}
\description{
Applies growth models to a \code{data.table} of forest inventory data, specifically using the
models developed by Maleki et al. (2022) for predicting stand-level increments over a
5-year period. The \code{ModelsAndParameters} list should contain the models and parameters
as described in the cited study.
}
\details{
The function iterates over the species specified in \code{nSpecies}, applying the corresponding
growth models from the \code{ModelsAndParameters} list to project the number of trees per hectare (N),
basal area (B), dominant height (H), and total standing volume (V) for each species group. These
models are based on the methodologies outlined in Maleki et al. (2022) and are intended for use
with Norwegian forest species groups.
}
\examples{
\dontrun{
## Assuming `Data` is your forest inventory data.table, `myMP` contains
## the models and parameters, and `mySSP` is the species vector:
## GrowthModels(Data = Data, ModelsAndParameters = myMP, nSpecies = mySSP)
}

}
\references{
Maleki et al. (2022). "Stand-level growth models for long-term projections of the main species groups in Norway".
\emph{Scandinavian Journal of Forest Research}, 37(1), 1-12.
\url{https://doi.org/10.1080/02827581.2022.2056632}
}
\seealso{
\code{\link{PixSim}}, \code{\link[=ModelsAndParameters]{ModelsAndParameters dataset}}
}
