% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-PixelTable.R
\docType{data}
\name{PixelTable}
\alias{PixelTable}
\title{PixelTable dataset}
\format{
A data frame with X rows and Y columns:
\describe{
\item{Column1}{Description of Column1}
\item{Column2}{Description of Column2}
...
}
}
\source{
Describe the source of the data, if applicable.
}
\usage{
PixelTable
}
\description{
A brief description of the dataset goes here. Explain what the dataset contains,
how it was collected, any important transformations or cleaning steps that were
performed, and any other relevant information.
}
\keyword{datasets}
