% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PixSim-package.R
\docType{package}
\name{PixSim-package}
\alias{PixSim}
\alias{PixSim-package}
\title{PixSim: Large-scale high-resolution forest simulations}
\description{
Perform large-scale high-resolution forest growth simulations.
}
\author{
\strong{Maintainer}: Nicolas Cattaneo \email{nicolas.cattaneo@nibio.no} (\href{https://orcid.org/0000-0002-1955-1174}{ORCID})

}
\keyword{internal}
