% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PixSim.R
\name{PixSim}
\alias{PixSim}
\title{Stand-Level Growth Simulation}
\usage{
PixSim(Data, Np, nSpecies, WriteOut = FALSE, LocalFldr = NULL, functions, ...)
}
\arguments{
\item{Data}{A \code{data.table} object representing a forest inventory derived from a Forest Resource map.
See \code{PixelTable} dataset in the package for table format example.}

\item{Np}{An integer indicating the number of projection time steps to simulate.}

\item{nSpecies}{A numeric vector specifying the species codes to simulate, corresponding to codes in \code{Data}.}

\item{WriteOut}{Logical; if \code{TRUE}, simulation results for each time step are saved to \code{LocalFldr}.}

\item{LocalFldr}{A character string specifying the path to a local folder where results should be written,
used if \code{WriteOut} is \code{TRUE}.}

\item{functions}{A named list of functions that define the growth models and other operations to be
applied during the simulation. At minimum, should contain a function with a forest growth model.
Refer to the \code{GrowthModels} function included in the package.}

\item{...}{Additional arguments to pass to the functions within the \code{functions} list.}
}
\description{
Performs stand-level growth simulations on a \code{PixelTable} created from Forest Resource Map rasters.
Users can refer to the \code{raster2pixeltable} documentation for details on creating a \code{PixelTable}.
The simulation projects the growth of forest stands over a specified number of time steps.
}
\details{
The \code{PixSim} function takes a \code{PixelTable}, the number of simulation periods (\code{Np}), the species to simulate (\code{nSpecies}),
and a list of functions (\code{functions}) that perform various steps in the simulation process. If \code{WriteOut} is \code{TRUE}, the
results of each simulation period are saved to disk at the location specified by \code{LocalFldr}. Extra parameters required
by the functions in \code{functions} can be passed through the \code{...} argument.
}
\examples{
\dontrun{
  PixelTableCopy <- data.table::copy(PixelTable)
  Functions <- list(GrowthModels = GrowthModels)
  myMM <- ModelsAndParameters[[1]]
  mySSP <- c(1, 2, 3)

  Fold <- tempfile()
  unlink(Fold, recursive = TRUE)
  dir.create(Fold)

  PixSim(Data = PixelTableCopy,
         Np = 15,
         nSpecies = mySSP,
         functions = Functions,
         WriteOut = TRUE,
         LocalFldr = Fold,
         ModelsAndParameters = myMM)

  Results <- list.files(Fold, full.names = TRUE)
  lapply(Results, fst::read_fst, as.data.table = TRUE)
}

}
