% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_SI_changes.R
\name{read_SI_changes}
\alias{read_SI_changes}
\title{Read Site Index Changes Data}
\usage{
read_SI_changes()
}
\value{
A \code{data.table} object with the site index changes data.
}
\description{
This function reads the \code{SI_changes.fst} file, which contains site index changes
for Norway's main forest species. These changes occur every five years from 2021 onwards
and match the coordinates in the \code{PixelTable} dataset. The data includes site index
values corresponding to various years, computed after Antón-Fernández et al. (2016)
\link{Climate-sensitive site index models for Norway;https://doi.org/10.1139/cjfr-2015-0155}.
}
\details{
The dataset is expected to be used in conjunction with \code{PixelTable} as it has the same
number of rows and the coordinates are aligned. Each column \code{SI_x} represents the site
index for a specific 5-year interval.
}
\examples{
# To read the SI_changes data as a data.table:
si_changes <- read_SI_changes()
# Now si_changes is a data.table object that can be used for analysis
head(si_changes)
}
