% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegFunction.R
\name{RegFunction}
\alias{RegFunction}
\title{Apply Regeneration Data to Initialize Pixels}
\usage{
RegFunction(Data, RegData)
}
\arguments{
\item{Data}{A \code{data.table} object representing the forest inventory to be simulated,
expected to be in the same format as the \code{PixelTable} dataset.}

\item{RegData}{A \code{data.table} with regeneration data for initializing simulation pixels,
structured as per the \code{RegData} dataset included in the package.}
}
\value{
The function modifies the \code{Data} table by reference; therefore, it does not return
anything. \code{Data} will have updated columns reflecting the initial state for
regeneration based on the \code{RegData} lookup table.
}
\description{
This function initializes the regeneration stage in the forest simulation by applying
regeneration data from a lookup table. It uses expert knowledge from the
\code{RegData} dataset to set the initial state for pixels that represent new growth following
clear-cutting.
}
\details{
The \code{RegFunction} is intended for use within the \code{PixSim} function as part of a larger
simulation workflow. After clear-cuts,
the \code{RegFunction} uses the \code{RegData} dataset to set the initial conditions for forest
regeneration. The function matches entries in \code{RegData} with the corresponding pixels in
\code{Data} by site index and species.
}
\examples{
\dontrun{
 ## PixelTableCopy <- copy(PixelTable)
 ## RegDataCopy <- copy(RegData)
 ## RegFunction(Data = PixelTableCopy, RegData = RegDataCopy)
 ## Now PixelTableCopy contains the initial regeneration data
}
}
\seealso{
\code{\link{PixSim}}, \code{\link[=RegData]{RegData dataset}}
}
