% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetAside.R
\name{SetAside}
\alias{SetAside}
\title{Set Aside Pixels for Environmental Considerations}
\usage{
SetAside(Data, SetAsideFile = NULL, SetAsidePercent = NULL)
}
\arguments{
\item{Data}{A \code{data.table} object representing the forest inventory, where certain pixels will be marked as restricted.}

\item{SetAsideFile}{A file path or a \code{data.table} object with a single column, aligned with \code{Data}, that indicates
which pixels to set aside with 1 (cannot be harvested) and 0 (can be harvested).}

\item{SetAsidePercent}{A numeric value representing the percentage of pixels within each forest stand to set aside.}
}
\value{
The function modifies \code{Data} by reference, adding an \code{envRestr} column that marks set-aside pixels.
It returns invisibly.
}
\description{
This function marks certain pixels in the forest inventory data table (\code{Data}) to be set aside from forest management,
which can be useful for preserving areas with environmental restrictions.
}
\details{
The function can take either a \code{SetAsideFile} and/or \code{SetAsidePercent} to determine which pixels to set aside. If a
\code{SetAsideFile} is provided, the function reads the file and marks the pixels accordingly. If \code{SetAsidePercent} is given,
it calculates the number of pixels to set aside within each stand based on the percentage. The pixels are then marked
with an \code{envRestr} flag in the \code{Data} table, where a value of 1 indicates a pixel set aside.
}
\examples{
\dontrun{
  ## Assuming `Data` is your forest inventory data.table
  ## Make a `SetAsideFile` file by labeling 20\% of the pixels in the forest inventory data:

  # library("data.table")
  # SetAsidePxls <- data.table(toSetAside = rep(0, nrow(Data)))
  # SetAsidePxls[sample(1:nrow(SetAsidePxls), nrow(SetAsidePxls) * 0.20), toSetAside := 1]
  # fst::save_fst(SetAsidePxls, "SetAsideFile.fst")

  SetAside(Data, SetAsideFile = "SetAsideFile.fst")
}
}
\seealso{
\code{\link{PixSim}}
}
